requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
itsdangerous==2.1.2
Jinja2==3.1.2
python-dotenv==1.0.0

PyPDF2==3.0.1

nltk==3.8.1
textblob==0.17.1
textstat==0.7.3
requests==2.31.0
matplotlib==3.8.2
numpy==1.26.4
seaborn==0.13.2
wordcloud==1.9.3

gunicorn==21.2.0
setuptools==69.5.1
